requirements.txt
streamlit
torch
transformers
Pillow
datasets